# Conflicts:
#	aes.c
#	aes.h
#	common.h
#	config.h
#	main.c
#	makefile
#	uart.c
#	uart.h

+++ main.c
@@ -6,91 +6,108 @@
 
 #include "config.h"
 #include "uart.h"
-#include "debugUART.h"
+#include "debug_uart.h"
 #include "common.h"
 #include "aes.h"
 
 static void get_command(APDU_command *cmd, APDU_response *response, int *errnum) {
-  int i;
+  uint8_t i;
   uint8_t in_buffer[6];
 
-  //reading command from uart. cmd is 5 bytes long
-  for(i=0; i<5; i++) {
-      uart_read_byte(&in_buffer[i]);
+  // reading command from uart. cmd is 5 bytes long
+  for (i = 0; i < 5; i++) {
+    uart_read_byte(&in_buffer[i]);
   }
 
   cmd->cla = in_buffer[0];
   cmd->ins = in_buffer[1];
-  cmd->p1 = in_buffer[2];
-  cmd->p2 = in_buffer[3];
-  cmd->p3 = in_buffer[4];
+  cmd->p1  = in_buffer[2];
+  cmd->p2  = in_buffer[3];
+  cmd->p3  = in_buffer[4];
 
   if (cmd->cla != CLA) {
-    //CLA invalid and aborting
+    // CLA invalid and aborting
     response->sw1 = 0x6e;
     response->sw2 = 0x00;
-    *errnum = ABORT_ERROR;
+    *errnum       = ABORT_ERROR;
   }
 }
 
-static void execute_command(APDU_command *cmd, APDU_response *response, int *errnum) {
+static void execute_command(APDU_command  *cmd,
+                            APDU_response *response,
+                            int           *errnum) {
   switch (cmd->ins) {
   case DECRYPT_DATA:
-    //sending ack
+
+    // sending ack
     uart_write_byte(&cmd->ins);
-    //read data from UART
+
+    // read data from UART
     uart_read(cmd->data, cmd->p3);
-    //decode data
+
+    // decode data
     aes_decrypt(cmd->data);
-    //send status bytes
+
+    // send status bytes
     response->sw1 = 0x61;
     response->sw2 = 0x10;
-    *errnum = NO_ERROR;
+    *errnum       = NO_ERROR;
     break;
+
   case GET_RESPONSE:
-    //sending ack
+
+    // sending ack
     uart_write_byte(&cmd->ins);
-    //sending decrypted data
+
+    // sending decrypted data
     uart_write(cmd->data, cmd->p3);
-    //send status bytes
+
+    // send status bytes
     response->sw1 = 0x90;
     response->sw2 = 0x00;
-    *errnum = NO_ERROR;
+    *errnum       = NO_ERROR;
     break;
+
   default:
-    //command not supported
+
+    // command not supported
     response->sw1 = 0x6d;
     response->sw2 = 0x00;
-    *errnum = WARNING_ERROR;
+    *errnum       = WARNING_ERROR;
   }
-  //writing status bytes
+
+  // writing status bytes
   uart_write_byte(&response->sw1);
   uart_write_byte(&response->sw2);
 }
 
 int main(void) {
   int errnum = 0;
-  APDU_command cmd;
+  APDU_command  cmd;
   APDU_response response;
-  uint8_t ATR[4] = {0x3B, 0x90, 0x11, 0x00};
+  uint8_t ATR[4] = { 0x3B, 0x90, 0x11, 0x00 };
 
-  //init functions here
+  // init functions here
   init_uart(&errnum);
   INIT_LED_PINS();
   #ifdef DEBUG
   init_debug_uart();
-  #endif
+  #endif /* ifdef DEBUG */
+
   // send ATR
   uart_write(ATR, 4);
-  while(1){
-    //getting command from uart
+
+  while (1) {
+    // getting command from uart
     get_command(&cmd, &response, &errnum);
+
     if (errnum == ABORT_ERROR) {
       uart_write_byte(&response.sw1);
       uart_write_byte(&response.sw2);
       break;
     }
-    //running command
+
+    // running command
     execute_command(&cmd, &response, &errnum);
 
     if (errnum == ABORT_ERROR) {
@@ -99,4 +116,3 @@ int main(void) {
   }
   return 0;
 }
-
